Fix the order in which completions (grouped by rule) are shown
Prior to this patch, completion rules would be evaluated in inverse order of
their definitions (e.g. in reverse order of the calls to `complete`), which led
to a number of completions being generated in an opposite order than intended.

This also makes a change to store completions associated with a command from an
`std::forward_list` to an `std::vector` collection. It does not seem that any of
the instances of `option_list_t` took advantage of any of the reasons for using
a list over a vector (we do not seem to persist pointers or iterators). This
should also improve performance by improving the cache friendliness and playing
nicer with the memory allocator.