Revert "maybe_t: make maybe_t<T> trivially copyable if T is"
This reverts commit 9d303a74e3197c82fe48171e9189912c9e0e2a22.
This reverts commit 0305c842e6056f926569e167afbef49f05c9c438.

9d303a7 broke 32-bit armhf builds for unknown reasons, specifically in
settings where a trivial copy of `maybe_t<int>` was performed. A caller
would pass a literal int in the place of a `maybe_t<int>` parameter and
the callee would see a populated `maybe_t` but with a value of `0`
rather than the actual value that was passed in. It was too painful to
debug to a resolution under qemu.