readline: make readline_cmd_t a class
In preparation for adding readline commands that have data associated
with them convert readline_cmd_t from an enum to a class. The enum is
renamed to readline_cmd_t::id_t. There are a lot of changes in this
commit but they all fall into one of three categories

 - rename readline_cmd_t to readline_cmd_t::id_t.
 - use the .id() instance method to access the enum member.
 - convert functions that take or return readline_cmd_t by value to
   use const references instead.