builtin commandline: use readline for text insertion
Instead of manipulating the command line buffer directly use a readline
command with associated data to insert text. This means that text
insertion by `commandline --insert` no-longer happens before any
`commandline -f` readline functions are executed.

The data for the readline function is stored in a discriminated
union. Ideally we'd use std::variant but that is not available in C++11
and the C++11 implementations I looked at were all huge and
complicated. Using a union requires constructors and assignment
operators that initialize the correct union member.

Before this change

    commandline -f backward-word
    commandline --insert "new text"

would insert "new text" at the current cursor position rather than at
the beginning of the current word because the text was inserted before
the `backward-word` command was executed.

This fixes and reinstates the test cases that were broken by the last
commit.