builtin commandline: use readline to set cursor position
This means that the cursor movement is no-longer executed out of order
when `commandline --cursor <pos>` is mixed with `commandline -f
<command>`

Before this change

    commandline -f begin-selection
    commandline --cursor 10
    commandline -f kill-selection end-selection

does not work as expected because the cursor is moved before any of the
readline functions are executed.

This change breaks a number of the commandline test cases that combine
`commandline --cursor` with `commandline --append/insert/replace`. They
are temporarily commented out, the next commit will make them pass
again.