builtin commandline: remove readline function hack
Now that `commandline --insert` is just another readline command it is
no-longer necessary to special case `begin-undo-group` and
`end-undo-group` when they are passed to `commandline -f`.

Before this change the `undo` command in

    commandline -f begin-undo-group
    commandline -f backward-delete-char backward-delete-char
    commandline -f end-undo-group undo

only undoes one deletion because the `begin-undo-group` and
`end-undo-group` commands are executed before the
`backwards-delete-char` and `undo` commands.

Fixes #3031