fish_git_prompt: initialize xcrun_db in background
This prevents fish appearing to hang on launch on macOS.

/usr/bin/git is a shim that finds Git in the Xcode or Command Line
Tools Developer dir. There is a cache file kept in TMPRDIR
that is cleared on reboot. The validity of the Xcode installation
is verified as well for security reasons, causing this to be
potentially really slow.

This runs a git no-op command in the background when we autoload
fish_git_prompt if the db file is missing on Darwin, with a
self-destructing callback that repaints the commandline when it's
done.

If the callback function exists, it means we're currently waiting
for the cold-start git command to finish, and do not print a
git_prompt.

Fixes #6625