Use xcrun --show-cache-path
Disassembling /usr/bin/xcrun demonstrated a couple things:

1) It is also a magic shim. It uses libxcselect.dylib from the
system dylib cache, and the actual command line tool `xcrun`
exists at e.g. /Applications/Xcode.app/Contents/Developer/usr/lib/libxcrun.dylib

2) The db file could exist elsewhere theoretically. The environment
variable `xcrun_db` is an absolute path to an alternate location.

3) Disassembling libxcrun.dylib revealed `xcrun` has a `--show-cache-path`
option that is not documented in the --help output, that outputs
the path correctly even if xcrun_db changed it. It outputs the right
path (even if the file does not exist, and no it does not need
the cache or cause the cache to be recreated.)

So, use this more reliable method.