history pager: allow filtering by multiple words
When using Control-R to search command history, I tend to not navigate the
pager. Instead I like to keep typing the query until my desired command is
on top.

This strategy is awkward when the query is part of a long string that is
shared by many commands.  Let's improve this scenario by interpreting the
history search string as glob pattern.